#1477. Fix type_inference_A01_t05.dart. Expect correct type

+++ LanguageFeatures/Records/type_inference_A01_t05.dart
@@ -41,5 +41,5 @@ class D<X extends Y, Y> extends C<Y> {}
 main() {
   (C<int>, {C<double> x}) r = (D(), x: D());
   Expect.equals(int, r.$0.typeArgument);
-  Expect.equals(int, r.x.typeArgument);
+  Expect.equals(double, r.x.typeArgument);
 }
